split logic and flask server
